Bug fixes and refactoring
-
