+++ Sources/Paywall/Misc/Extensions/UIApplication+ActiveWindow.swift
@@ -10,11 +10,7 @@ import UIKit
 
 extension UIApplication {
   var activeWindow: UIWindow? {
-    if #available(iOS 13.0, *) {
-      let windows = UIApplication.shared.connectedScenes.flatMap { ($0 as? UIWindowScene)?.windows ?? [] }
-      return windows.first { $0.isKeyWindow } ?? windows.first
-    } else {
-      return UIApplication.shared.keyWindow
-    }
+    let windows = UIApplication.shared.connectedScenes.flatMap { ($0 as? UIWindowScene)?.windows ?? [] }
+    return windows.first { $0.isKeyWindow } ?? windows.first
   }
 }
