+++ Sources/Paywall/Misc/Extensions/UIWindow+Landscape.swift
@@ -9,14 +9,10 @@ import UIKit
 
 extension UIWindow {
   static var isLandscape: Bool {
-    if #available(iOS 13.0, *) {
-      return UIApplication.shared.windows
-        .first?
-        .windowScene?
-        .interfaceOrientation
-        .isLandscape ?? false
-    } else {
-      return UIApplication.shared.statusBarOrientation.isLandscape
-    }
+    return UIApplication.shared.windows
+      .first?
+      .windowScene?
+      .interfaceOrientation
+      .isLandscape ?? false
   }
 }
