+++ Sources/Paywall/Paywall/Paywall Manager/PaywallCache.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 
-@MainActor
 final class PaywallCache {
   func getPaywall(
     withIdentifier identifier: String?
