+++ Sources/Paywall/Paywall/Paywall Presentation/InternalPresentationLogic.swift
@@ -38,4 +38,20 @@ enum InternalPresentationLogic {
 
     return checkSubscriptionStatus()
   }
+
+  static func presentationError(
+    domain: String,
+    code: Int,
+    title: String,
+    value: String
+  ) -> NSError {
+    let userInfo: [String: Any] = [
+      NSLocalizedDescriptionKey: NSLocalizedString(title, value: value, comment: "")
+    ]
+    return NSError(
+      domain: domain,
+      code: code,
+      userInfo: userInfo
+    )
+  }
 }
