+++ Sources/Paywall/Paywall/Paywall Presentation/Operators/AwaitIdentityOperator.swift
@@ -10,9 +10,13 @@ import Combine
 
 extension AnyPublisher where Output == PaywallPresentationRequest, Failure == Error {
   func awaitIdentity() -> AnyPublisher<PaywallPresentationRequest, Failure> {
-    self
+    let hasIdentityPublisher = IdentityManager.shared.hasIdentity
+      .filter { $0 == true }
+      .setFailureType(to: Error.self)
+
+    return self
       .subscribe(on: DispatchQueue.global(qos: .userInitiated))
-      .zip(IdentityManager.shared.$hasIdentity.isTrue()) { request, hasIdentity in
+      .zip(hasIdentityPublisher) { request, hasIdentity in
         return request
       }
       .eraseToAnyPublisher()
