+++ Sources/Paywall/Paywall/Paywall Presentation/Operators/TriggerOutcomeOperator.swift
