+++ Sources/Paywall/Paywall/Paywall View Controller/WebEventHandler.swift
@@ -250,11 +250,6 @@ final class WebEventHandler: WebEventDelegate {
     if Paywall.options.isGameControllerEnabled {
       return
     }
-
-    if #available(iOS 13.0, *) {
-      UIImpactFeedbackGenerator().impactOccurred(intensity: 0.7)
-    } else {
-      UIImpactFeedbackGenerator().impactOccurred()
-    }
+    UIImpactFeedbackGenerator().impactOccurred(intensity: 0.7)
   }
 }
